Revert "Revert "Merge branch 'master' of https://github.com/patarshn/newsid""
This reverts commit 023157aa09ca36921e2738adb60650c6aba3c553.

+++ application/controllers/admin/Buku_data_peraturan_bpd.php
@@ -6,6 +6,7 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
     private $_table = 'buku_data_peraturan_bpd';
     private $_folder = 'buku_data_peraturan_bpd';
     private $_mainTitle = 'Buku Data Peraturan/Keputusan BPD';
+    private $_docxName = 'buku_data_peraturan_bpd.docx';
 
     function __construct() {
         parent::__construct();
@@ -16,20 +17,20 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
 
     function rulesStore() {
         return [
-            ['field' => 'no_dan_tgl_peraturan','label' => 'Nomor dan Tanggal Peraturan/Keputusan BPD', 'rules' => 'required'],
+            ['field' => 'no_peraturan','label' => 'Nomor Peraturan/Keputusan BPD', 'rules' => 'required'],
+            ['field' => 'tgl_peraturan','label' => 'Tanggal Peraturan/Keputusan BPD', 'rules' => 'required'],
             ['field' => 'tentang','label' => 'Tentang', 'rules' => 'required'],
             ['field' => 'uraian_singkat','label' => 'Uraian Singkat', 'rules' => 'required'],
-            ['field' => 'ket','label' => 'Keterangan', 'rules' => 'required'],
         ];
     }
 
     function rulesUpdate() {
         return [
             ['field' => 'id','label' => 'id', 'rules' => 'required'],
-            ['field' => 'no_dan_tgl_peraturan','label' => 'Nomor dan Tanggal Peraturan/Keputusan BPD', 'rules' => 'required'],
+            ['field' => 'no_peraturan','label' => 'Nomor Peraturan/Keputusan BPD', 'rules' => 'required'],
+            ['field' => 'tgl_peraturan','label' => 'Tanggal Peraturan/Keputusan BPD', 'rules' => 'required'],
             ['field' => 'tentang','label' => 'Tentang', 'rules' => 'required'],
             ['field' => 'uraian_singkat','label' => 'Uraian Singkat', 'rules' => 'required'],
-            ['field' => 'ket','label' => 'Keterangan', 'rules' => 'required'],
         ];
     }
 
@@ -94,10 +95,9 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
                 if(!$berkas){
-                    echo $this->upload->display_errors();
                     $callback = array(
                         'status' => 'error',
-                        'message' => 'Mohon Maaf, file gagal diupload',
+                        'message' => $this->upload->display_errors(),
                     );
                     echo json_encode($callback);
                     exit;
@@ -109,7 +109,8 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
             }
             $_POST = $this->input->post();
             $data = array(
-                'no_dan_tgl_peraturan' => $_POST['no_dan_tgl_peraturan'],
+                'no_peraturan' => $_POST['no_peraturan'],
+                'tgl_peraturan' => $_POST['tgl_peraturan'],
                 'tentang' => $_POST['tentang'],
                 'uraian_singkat' => $_POST['uraian_singkat'],
                 'ket' => $_POST['ket'],
@@ -182,6 +183,14 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
             //jika ada file yang baru
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
+                if(!$berkas){
+                    $callback = array(
+                        'status' => 'error',
+                        'message' => $this->upload->display_errors(),
+                    );
+                    echo json_encode($callback);
+                    exit;
+                }
                 $berkas_lama = $this->destroy_file($where);
             }
 
@@ -190,7 +199,8 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
                 $berkas = $_POST["old_file"];
             }
             $data = array(
-                'no_dan_tgl_peraturan' => $_POST['no_dan_tgl_peraturan'],
+                'no_peraturan' => $_POST['no_peraturan'],
+                'tgl_peraturan' => $_POST['tgl_peraturan'],
                 'tentang' => $_POST['tentang'],
                 'uraian_singkat' => $_POST['uraian_singkat'],
                 'ket' => $_POST['ket'],
@@ -261,7 +271,7 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
 
             if($this->Main_m->destroy($this->_table,$where)){
                 
-                $this->session->set_flashdata('success_message', 'Delete form berhasil, terimakasih');
+                $this->session->set_flashdata('success_message', 'Hapus form berhasil, terimakasih');
                 $callback = array(
                     'status' => 'success',
                     'message' => 'Data berhasil dihapus',
@@ -269,7 +279,7 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
                 );
             }
             else{
-                $this->session->set_flashdata('error_message', 'Mohon maaf, delete form gagal');
+                $this->session->set_flashdata('error_message', 'Mohon maaf, hapus form gagal');
                 $callback = array(
                     'status' => 'error',
                     'message' => 'Mohon Maaf, Pengisian form gagal',
@@ -404,7 +414,7 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
             return $this->upload->data("file_name");
         }
         else{
-            echo $this->upload->display_errors();
+            return false;
         }    
     }
 
@@ -416,12 +426,51 @@ class Buku_data_peraturan_bpd extends Admin_Controller {
                 return true;
             }
 
+            if (!file_exists(FCPATH."administrasilainnya/" .$this->_folder."/".$b_id->berkas)){
+                return true;
+            }
+
             if (!unlink(FCPATH."administrasilainnya/".$this->_folder."/".$b_id->berkas)) {
                 return false;
             }
             
         }
         return true;
+    }
+
+    public function cetak(){
+        $data = $this->Main_m->get($this->_table,null)->result();
+        $today = date('Y-m-d');
+        $phpWord = new \PhpOffice\PhpWord\PhpWord();
+        $templateProcessor = $phpWord->loadTemplate('./assets/buku_adm_lain/'.$this->_docxName);
+        $values = array();
+        $no = 1;
+        foreach($data as $d){
+            $subvalues = array(
+                'no' => $no++,
+                'no_peraturan' => $d->no_peraturan,
+                'tgl_peraturan' => $d->tgl_peraturan,
+                'tentang' => $d->tentang,
+                'uraian_singkat' => $d->uraian_singkat,
+                'ket'=> $d->ket
+            );
+            $values[] = $subvalues;
+        }
+        $templateProcessor->cloneRowAndSetValues('no', $values);
+        $temp_filename = $this->_docxName;
+        $templateProcessor->saveAs($temp_filename);
+        header('Content-Description: File Transfer');
+        header('Content-Type: application/octet-stream');
+        header('Content-Disposition: attachment; filename='.$temp_filename);
+        header('Content-Transfer-Encoding: binary');
+        header('Expires: 0');
+        header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
+        header('Pragma: public');
+        header('Content-Length: ' . filesize($temp_filename));
+        flush();
+        readfile($temp_filename);
+        unlink($temp_filename);
+        exit;
     }
 }
 ?>
