+++ application/controllers/admin/Buku_lemdes_berdes.php
@@ -1,11 +1,16 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+use PhpOffice\PhpSpreadsheet\IOFactory;
+use PhpOffice\PhpSpreadsheet\Spreadsheet;
+use PhpOffice\PhpSpreadsheet\Writer\Xls;
+
 class Buku_lemdes_berdes extends Admin_Controller {
 
     private $_table = 'buku_lemdes_berdes';
     private $_folder = 'buku_lemdes_berdes';
     private $_mainTitle = 'Buku Lembaran Desa dan Berita Desa';
+    private $_exelName = 'buku_lemdes_berdes.xls';
 
     function __construct() {
         parent::__construct();
@@ -17,7 +22,8 @@ class Buku_lemdes_berdes extends Admin_Controller {
     function rulesStore() {
         return [
             ['field' => 'jns_peraturan_desa','label' => 'jns_peraturan_desa', 'rules' => 'required'],
-            ['field' => 'no_tgl_ditetapkan','label' => 'no_tgl_ditetapkan', 'rules' => 'required'],
+            ['field' => 'no_ditetapkan','label' => 'no_ditetapkan', 'rules' => 'required'],
+            ['field' => 'tgl_ditetapkan','label' => 'tgl_ditetapkan', 'rules' => 'required'],
             ['field' => 'tentang','label' => 'tentang', 'rules' => 'required'],
             ['field' => 'tgl_diundangkan','label' => 'tgl_diundangkan', 'rules' => 'required'],
             ['field' => 'no_diundangkan','label' => 'no_diundangkan', 'rules' => 'required'],
@@ -28,7 +34,8 @@ class Buku_lemdes_berdes extends Admin_Controller {
     function rulesUpdate() {
         return [
             ['field' => 'jns_peraturan_desa','label' => 'jns_peraturan_desa', 'rules' => 'required'],
-            ['field' => 'no_tgl_ditetapkan','label' => 'no_tgl_ditetapkan', 'rules' => 'required'],
+            ['field' => 'no_ditetapkan','label' => 'no_ditetapkan', 'rules' => 'required'],
+            ['field' => 'tgl_ditetapkan','label' => 'tgl_ditetapkan', 'rules' => 'required'],
             ['field' => 'tentang','label' => 'tentang', 'rules' => 'required'],
             ['field' => 'tgl_diundangkan','label' => 'tgl_diundangkan', 'rules' => 'required'],
             ['field' => 'no_diundangkan','label' => 'no_diundangkan', 'rules' => 'required'],
@@ -97,10 +104,10 @@ class Buku_lemdes_berdes extends Admin_Controller {
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
                 if(!$berkas){
-                    echo $this->upload->display_errors();
+                    
                     $callback = array(
                         'status' => 'error',
-                        'message' => 'Mohon Maaf, file gagal diupload',
+                        'message' => $this->upload->display_errors(),
                     );
                     echo json_encode($callback);
                     exit;
@@ -114,7 +121,8 @@ class Buku_lemdes_berdes extends Admin_Controller {
                 $data = array(
                     'jns_peraturan_desa' => $_POST['jns_peraturan_desa'],
                     'tentang' => $_POST['tentang'],
-                    'no_tgl_ditetapkan' => $_POST['no_tgl_ditetapkan'],
+                    'no_ditetapkan' => $_POST['no_ditetapkan'],
+                    'tgl_ditetapkan' => $_POST['tgl_ditetapkan'],
                     'tgl_diundangkan' => $_POST['tgl_diundangkan'],
                     'no_diundangkan' => $_POST['no_diundangkan'],
                     'ket' => $_POST['ket'],
@@ -189,6 +197,14 @@ class Buku_lemdes_berdes extends Admin_Controller {
             //jika ada file yang baru
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
+                if(!$berkas){
+                    $callback = array(
+                        'status' => 'error',
+                        'message' => $this->upload->display_errors(),
+                    );
+                    echo json_encode($callback);
+                    exit;
+                }
                 $berkas_lama = $this->destroy_file($where);
             }
 
@@ -200,7 +216,8 @@ class Buku_lemdes_berdes extends Admin_Controller {
             $data = array(
                 'jns_peraturan_desa' => $_POST['jns_peraturan_desa'],
                 'tentang' => $_POST['tentang'],
-                'no_tgl_ditetapkan' => $_POST['no_tgl_ditetapkan'],
+                'no_ditetapkan' => $_POST['no_ditetapkan'],
+                'tgl_ditetapkan' => $_POST['tgl_ditetapkan'],
                 'tgl_diundangkan' => $_POST['tgl_diundangkan'],
                 'no_diundangkan' => $_POST['no_diundangkan'],
                 'ket' => $_POST['ket'],
@@ -414,7 +431,7 @@ class Buku_lemdes_berdes extends Admin_Controller {
             return $this->upload->data("file_name");
         }
         else{
-            echo $this->upload->display_errors();
+            return false;
         }    
     }
 
@@ -425,6 +442,10 @@ class Buku_lemdes_berdes extends Admin_Controller {
             if(empty($b_id->berkas)){
                 return true;
             }
+
+            if (!file_exists($b_id->berkas)){
+                return true;
+            }
             
             if (!unlink(FCPATH."uploads/".$this->_folder."/".$b_id->berkas)) {
                 return false;
@@ -432,6 +453,68 @@ class Buku_lemdes_berdes extends Admin_Controller {
             
         }
         return true;
+    }
+
+    public function cetak(){
+        $reader = IOFactory::createReader('Xls');
+        $spreadsheet = $reader->load('./assets/buku_adm_umum/'.$this->_exelName);
+        $data = $this->Main_m->get($this->_table,null)->result();
+        $values = array();
+        $i = 0;
+        $no = 1;
+        foreach($data as $d){
+            $subvalues = array(
+                $no++,
+                $d->jns_peraturan_desa,
+                $d->no_ditetapkan.", ".$d->tgl_ditetapkan,
+                $d->tentang,
+                $d->tgl_diundangkan,
+                $d->no_diundangkan,
+                $d->ket
+            );
+            $values[] = $subvalues;
+            $i++;
+        }
+
+        $sheet = $spreadsheet->getActiveSheet();
+        $sheet->fromArray(
+            $values,
+            NULL,
+            'A8'
+        );
+
+        $styleArray = [
+            'borders' => [
+                'allBorders' => [
+                    'borderStyle' => \PhpOffice\PhpSpreadsheet\Style\Border::BORDER_THIN,  
+                ],
+            ],
+        ];
+
+        $i = $i + 7;
+
+        $sheet->getStyle('A8:G'.$i)->applyFromArray($styleArray);
+        $sheet->getStyle('A8:G'.$i)->getAlignment()->setWrapText(true);
+        $sheet->getStyle('A8:G'.$i)->getAlignment()->setHorizontal(\PhpOffice\PhpSpreadsheet\Style\Alignment::HORIZONTAL_LEFT);
+        $sheet->getStyle('A8:G'.$i)->getAlignment()->setVertical(\PhpOffice\PhpSpreadsheet\Style\Alignment::VERTICAL_CENTER);
+        // foreach(range('A7','J') as $columnID) {
+        //     $sheet->getColumnDimension($columnID)->setAutoSize(true);
+        // }
+        for($r = 8;$r <= $i;$r++){
+            $sheet->getRowDimension((string)$r)->setRowHeight(-1);
+        }
+        $writer = new Xls($spreadsheet);
+
+        $filename = $this->_exelName;
+
+        header('Content-Description: File Transfer');
+        header('Content-Type: application/vnd.ms-excel');
+        header('Content-Disposition: attachment; filename='.$filename);
+        header('Content-Transfer-Encoding: binary');
+        header('Expires: 0');
+        header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
+        header('Pragma: public');
+        $writer->save('php://output');
     }
 }
 ?>
