+++ application/controllers/admin/Buku_tnh_desa.php
@@ -1,11 +1,16 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+use PhpOffice\PhpSpreadsheet\IOFactory;
+use PhpOffice\PhpSpreadsheet\Spreadsheet;
+use PhpOffice\PhpSpreadsheet\Writer\Xls;
+
 class Buku_tnh_desa extends Admin_Controller {
 
     private $_table = 'buku_tnh_desa';
     private $_folder = 'buku_tnh_desa';
     private $_mainTitle = 'Buku Tanah Di Desa';
+    private $_exelName = 'buku_tnh_desa.xls';
 
     function __construct() {
         parent::__construct();
@@ -133,10 +138,10 @@ class Buku_tnh_desa extends Admin_Controller {
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
                 if(!$berkas){
-                    echo $this->upload->display_errors();
+                    
                     $callback = array(
                         'status' => 'error',
-                        'message' => 'Mohon Maaf, file gagal diupload',
+                        'message' => $this->upload->display_errors(),
                     );
                     echo json_encode($callback);
                     exit;
@@ -244,6 +249,14 @@ class Buku_tnh_desa extends Admin_Controller {
             //jika ada file yang baru
             if(!empty($_FILES["berkas"]["name"])){
                 $berkas = $this->upload_file();
+                if(!$berkas){
+                    $callback = array(
+                        'status' => 'error',
+                        'message' => $this->upload->display_errors(),
+                    );
+                    echo json_encode($callback);
+                    exit;
+                }
                 $berkas_lama = $this->destroy_file($where);
             }
 
@@ -488,7 +501,7 @@ class Buku_tnh_desa extends Admin_Controller {
             return $this->upload->data("file_name");
         }
         else{
-            echo $this->upload->display_errors();
+            return false;
         }    
     }
 
@@ -500,12 +513,97 @@ class Buku_tnh_desa extends Admin_Controller {
                 return true;
             }
 
+            if (!file_exists(FCPATH."uploads/".$this->_folder."/".$b_id->berkas)){
+                return true;
+            }
+
             if (!unlink(FCPATH."uploads/".$this->_folder."/".$b_id->berkas)) {
                 return false;
             }
             
         }
         return true;
+    }
+
+    public function cetak(){
+        $reader = IOFactory::createReader('Xls');
+        $spreadsheet = $reader->load('./assets/buku_adm_umum/'.$this->_exelName);
+        $data = $this->Main_m->get($this->_table,null)->result();
+        $values = array();
+        $i = 0;
+        $no = 1;
+        foreach($data as $d){
+            $subvalues = array(
+                $no++,
+                $d->nama_perorangan_bdn_hkm,
+                $d->jml,
+                $d->sdh_serti_hm,
+                $d->sdh_serti_hgb,
+                $d->sdh_serti_hp,
+                $d->sdh_serti_hgu,
+                $d->sdh_serti_hpl,
+                $d->blm_serti_ma,
+                $d->blm_serti_vi,
+                $d->blm_serti_tn,
+                $d->non_pertanian_perumahan,
+                $d->non_pertanian_dagang_jasa,
+                $d->non_pertanian_kantor,
+                $d->non_pertanian_industri,
+                $d->non_pertanian_fasilitas_umum,
+                $d->pertanian_sawah,
+                $d->pertanian_tegalan,
+                $d->pertanian_kebun,
+                $d-> pertanian_ternak,
+                $d->pertanian_hutan,
+                $d->pertanian_hutan_lindung,
+                $d->pertanian_mutasi,
+                $d->pertanian_tanah_kosong,
+                $d->pertanian_lain,
+                $d->ket
+            );
+            $values[] = $subvalues;
+            $i++;
+        }
+
+        $sheet = $spreadsheet->getActiveSheet();
+        $sheet->fromArray(
+            $values,
+            NULL,
+            'A9'
+        );
+
+        $styleArray = [
+            'borders' => [
+                'allBorders' => [
+                    'borderStyle' => \PhpOffice\PhpSpreadsheet\Style\Border::BORDER_THIN,  
+                ],
+            ],
+        ];
+
+        $i = $i + 8;
+
+        $sheet->getStyle('A9:Z'.$i)->applyFromArray($styleArray);
+        $sheet->getStyle('A9:Z'.$i)->getAlignment()->setWrapText(true);
+        $sheet->getStyle('A9:Z'.$i)->getAlignment()->setHorizontal(\PhpOffice\PhpSpreadsheet\Style\Alignment::HORIZONTAL_LEFT);
+        $sheet->getStyle('A9:Z'.$i)->getAlignment()->setVertical(\PhpOffice\PhpSpreadsheet\Style\Alignment::VERTICAL_CENTER);
+        // foreach(range('A7','J') as $columnID) {
+        //     $sheet->getColumnDimension($columnID)->setAutoSize(true);
+        // }
+        for($r = 9;$r <= $i;$r++){
+            $sheet->getRowDimension((string)$r)->setRowHeight(-1);
+        }
+        $writer = new Xls($spreadsheet);
+
+        $filename = $this->_exelName;
+
+        header('Content-Description: File Transfer');
+        header('Content-Type: application/vnd.ms-excel');
+        header('Content-Disposition: attachment; filename='.$filename);
+        header('Content-Transfer-Encoding: binary');
+        header('Expires: 0');
+        header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
+        header('Pragma: public');
+        $writer->save('php://output');
     }
 }
 ?>
